fan_service : Fix fan_service bug (unable to kick watchdog)
Summary:
fan_service : Fix fan_service bug (unable to kick watchdog)
The rootcause of this issue is that, watchdog value configuration is empty as the relevant line in the configuration file was not parsed.
We did not define "value" keyword, and tried use it, which caused the parsing of this line to fail, which in turn caused watchdog fail to kick.

Reviewed By: harshitgulati18, binhuang00

Differential Revision: D39440886

fbshipit-source-id: b2a537ae6a06d166691510d08ac123f4b162645a