make hw tests to use desired loopback mode
Summary:
As titled, move hw tests to use desired loopback mode instead of harcoding it.

This is done so that we can leverage running hw tests on platforms with external phy.

Configure npu in loopback none but xphy in loopback mode so that the packet reaches the phy and then loops back.

Reviewed By: nivinl

Differential Revision: D39806705

fbshipit-source-id: 1e372a422ab1c4ae831729b389d121206640fa1b