Fix race b/w shutdown and route advertisement
Summary:
Currently there is a race b/w route advertisement and pcapMgr destruction that can be triggered 2 ways
1. RouteAdvertiser destructor queues one last advertisement on bg thread but does not wait for it to finish. We then go on and start destorying rest of SwSwitch including pcapMgr and end up seeing a crash when we actually try to send a pkt out P533585231.
2. More likely race is the timeout firing after we destroy pcapMgr but before we stop bg thread, and we get a similar result.

Cancelling timeout in stop() helps, but since we don't wait for stop() to finish before destorying RouteAdvertiser (and subsequently Ipv6Handler and pcapMgr), both races would still be possible. So simple soln here is to make sure to keep pcapMgr until after bg thread is killed. Other soln would be to wait for stop() to finish before finishing IPPv6Handler destructor, but not sure if that's as safe

Differential Revision: D39950313

fbshipit-source-id: 28f9bd8c9dd4542adaa508fd3e9e7ecdee8f9c7b