Stop traffic at the end of traffic test
Summary:
With SAI code using the new SDK, in tearDown flow, some of the
configurations get unset and for these configs, its needed that there is no
traffic running in the system. If there is traffic, the unconfig flow will
timeout in SDK waiting for traffic to stop. Hence modifying the test to have a
link flap at the end of verify to stop the traffic which can let the TearDown
work fine.

Worked with Broadcom in CS00012264174.

Reviewed By: zechengh09

Differential Revision: D39874739

fbshipit-source-id: 99a9da57cbdd95b54e41ed4cd7f266dab61fc41f