Change flexcounter programming model for EM stats to scale better
Summary:
The EM stats code is shared with acl stats. This leads to a scaling challenge in TH4 which restricts stats to 32.

existing acl stats flexcounter programming model:
- Allocate one flexcounteraction per stat
- attach stat to acl rule

The problem here is that we cannot have more than 32 flexcounteractionss.

new programming model for EM stats
- Allocate one flexcounter globally
- attach same flexcounter to all EM rules but using different action indexs
- eg: with 100 EM entries, we will allocate one flexcounter with 100 action indexes and attach each flexcounter action index to each EM entry

sorry about the long change list but the changes are straight forward

summary of changes:
- add a stats action index to objects and apis that use stats handle
- refactor acl stats code such that it can handle dedicated flex counter action for acl and also a shared flex counter action for EM

Reviewed By: rskumaran-fb

Differential Revision: D39826379

fbshipit-source-id: 4ef0c3e9a51b9eb6c5e9dce758fd1bfbd4337619