Configure RxEqualization setting for SFF optics if specified in config
Summary: Currently, we have a special case for yamp in qsfp service code, and we override the rx equalization setting if the platform type is yamp. Instead, we should be using qsfp config for this purpose. QSFP config for yamp specifies the settings that need to be overridden. In this diff, I change the QsfpModule code to honor the settings defined in the config and program them.

Differential Revision: D38986219

fbshipit-source-id: de0d24a19ec520d5da2781b38824d6ae29689365