Fix readRegister function in wedge_qsfp_util
Summary:
readRegister calls doReadRegViaService which returns a ReadResponse. ReadResponse stores a IOBuf that stores a pointer to the underlying buffer. When IOBuf gets out of scope, the underlying buffer is released as well.

Now, in readRegister we were storing the pointer to the underlying buffer in another bufPtr (which the user passed to it). The underlying data gets cleaned up at the end of readRegister so the callers of readRegister see garbage in *buf.

This diff fixes the problem by assuring the data is copied to the required buffer before it gets cleaned up.

Reviewed By: rajank7

Differential Revision: D40289156

fbshipit-source-id: 916ce49182cc9a3a5d8d91840475c027e0f2871f