Merge stream and client threads
Summary:
In the event of Fsdb crash, to properly get immediate notifications that stream has shutdown I think client and stream need to happen on the same thread. This also raised a tsan issue since the exception is created on the client thread but then accessed from the stream thread https://fb.workplace.com/groups/thriftusers/posts/1045771566132255
As a bonus, merging also gives more control to the client on the threading model. SA usecase is to share the same thread pool across different subscriptions

Differential Revision: D40291656

fbshipit-source-id: 9255a8250f3631e16649417acd8abd5dc69fc7d0