Towards supporting delta functions on non NodeMap containers
Summary:
With Thrift migration we will need to support
delta functions for non NodeMap containers. A immediate
requirement is for std::map<string, NeighborEntry> tables.

Here we will store pointer to old and new map values in
delta. So as a first step add a type alias for how the
the Node value is store. For NodeMaps it will be inside
a shared_ptr while for non NodeMaps it will just be
a raw pointer.

Differential Revision:
D40462304

LaMa Project: L1125642

fbshipit-source-id: b673c5a8a5f1edebb3ab6c6dc8c60a7a0abc45da