Allow for missing vlan->interface mapping in EcmpSetupHelper
Summary:
For bcm link tests, in wedge100 config we are creating
vlan 4094 and associating port 2 with it. But no interface for
this vlan is created. This seems like a config gen oversight,
but make EcmpSetupHelper smart enough to work around this.
Note that EcmpSetupHelper can't really do anything with such
vlans anyways. Since it needs intf + ip addresses to be
able to create valid nhops.

Reviewed By: harshitgulati18, msomasundaran

Differential Revision:
D40542576

LaMa Project: L1125642

fbshipit-source-id: 99bc8a64094af3b42def3849bcaf09777a9b8432