fan_service : On Darwin, reduce boost_pwm to 60%
Summary:
[fboss/platform] fan_service : On Darwin, reduce boost_pwm to 60%
On Darwin PWM boost will kick in when there is no data from qsfp_service for more than 90 seconds.
On qsfp_service, the maximum pwm in the worst case is 60%. So reducing it to 60%.
According to the fan_service algorithm, if any sensor triggers pwm to be more than 60%, the higher value will still override this boost pwm value. So it's safe to reduce this value to 60%.

Reviewed By: jasmeetbagga

Differential Revision: D40532833

fbshipit-source-id: 53ce657cb2ddaed1ff044c94252d773fc86cfe13