Add ndp, arp tables to Interface state
Summary:
With VOQ switches, RIFs are no longer tied to
VLANs. In particular we create RIFs that are of type port
and are variously associated physical port (local) or
sysports (remote). For this use case, we will have
to move neighbor entries to be tied with RIFs.
Start doing so in switch state

Reviewed By: peygar

Differential Revision: D40121747

fbshipit-source-id: f0190d2e5983d8e932da624e4781aaf445c1d513