Set Rx and Tx PFC attributes seperately if needed
Summary:
Adding the change for completeness to make sure we can set Rx and Tx
PFC differently if needed to be on par with native implementation. This suport
is currently available only in Broadcom, but in prod, out config has both Rx
and Tx either turned on or off together and this flow might not be hit, but
tests would cover the combinations.

Reviewed By: zechengh09

Differential Revision: D40525123

fbshipit-source-id: 9be782b06d146135142934c265dde2b0908dd087