a multi node test for routing loop
Summary:
routing loop test mimicking  scenario as seen in https://fburl.com/q2rm8lmh

this test produces 100% traffic loss (or 100% in-discards) on DUT.

a routing loop is established with non-default prefix route and traffic with destination ip matching that prefix is injected from remote device (UU). On DUT (DU), 100% traffic loss is observed. The hash seeds are set to exact values as what are seen on UU and DU on the setup and packet injected are of same flow signature as they were injected on DU-UU routing loop setup with 100% traffic loss.

Reviewed By: msomasundaran

Differential Revision: D40731178

fbshipit-source-id: 32eeb75e43172837b2e118019da8735f8a8d52d7