Program remote neighbors with fixed encap idx
Summary:
As titled. We need this to bootstrap inband
connectivity. For now I am making a simplifying assumption
that all IN nodes in a DSF cluster have the same ASIC.
This is fine for the forseable future. If/when this
changes,
- We can align reserved encap idx range on the new ASIC to
be the same as previous ASIC
- Encode asic type in DSF node config

Reviewed By: shri-khare

Differential Revision:
D40882530

LaMa Project: L1125642

fbshipit-source-id: 6aa2a08bacb0d04d5ab72e421dac79891145fd0a