Add exception stack trace to MKA reload config
Summary:
This is to help debugging D39916600 where the only output is this:
```
unknown file: Failure
C++ exception with description "TTransportException: Timed out" thrown in the test body.
```
It's very hard to tell who threw the exception.

Only covering the reload config I need for now. But with this tool, we print the original stack trace, not the one in the catch block. See https://www.internalfb.com/code/fbsource/fbcode/folly/experimental/exception_tracer/README.md for details. So later on we can clean this up and wrap around the entire test body to print stack trace on any arbitrary exception. Gtest does not allow us to insert a wrapper. But we could potentially put the wrapper in `verifyAcrossWarmBoots()` and almost all tests will benefit. The exception is rethrown so there's no change in behavior.

Reviewed By: birdsoup

Differential Revision:
D40879509

LaMa Project: L1125642

fbshipit-source-id: b59c78eb956afb50525d9931fd7ce2549e45c421