Fix OSS build
Summary:
For some reason, Folly::folly_exception_tracer* is not available in OSS.
Need to check with folly team for this.

Since exception tracer is really used in closed source MacSec test only as in
diff D40879509, excluding its function template in OSS compilation for now until
we have a better fix.

Reviewed By: shri-khare

Differential Revision: D41048172

fbshipit-source-id: ff4bab809b8f0f14a43e11834e46c1df63f62c73