Sandia XPHY host side -> NPU IPHY tuning function
Summary:
As per our port bring up state machine, the NPU IPHY port gets programmed first and then XPHY port gets programmed. So in Sandia we observed that the NPU port Rx is locking to the wrong signal level coming from XPHY host Tx.
To make the NPU Rx adapt to the right signal from XPHY host Tx, we need to flap the XPHY host Tx after the XPHY port is created.
The function in this diff takes care of that

Reviewed By: harshitgulati18

Differential Revision: D41007165

fbshipit-source-id: bebefcfe6744134c1a62141445fcd846c8166f19