thrift struct fields to be qualified with derived class template parameter
Summary:
this refactoring is being done to make sure there can be multiple derived classes for same thrift object.

for a given thrift type, a derived class can inject its own type but this mapping is not available to container objects such as struct.

this refactoring will allow container object to specify mapping it desires for thrift object, when thrift object is member of container object

example is map<string, neighbor entry fields>. this needs to be modeled both as arp table and ndp table.

Reviewed By: jasmeetbagga

Differential Revision:
D41063212

LaMa Project: L1125642

fbshipit-source-id: 42bb6ce8c60810def6ffc248395696ef4f1120c1