fix broken MacLearningTest.l2EntryFlap on darwin switch
Summary: MacLearningTest.l2EntryFlap is failing on darwin switch because learned L2 entry does not have expected vlan. This turns out to be a bug in test, where we were assuming getCabledPorts()[0] and getCabledPorts()[1] are connected. It seems like this assumption is no longer true. So, use getConnectedPairs() to explicitly get a pair of connected ports.

Reviewed By: harshitgulati18

Differential Revision: D41140122

fbshipit-source-id: 45874f13ee7b11b27d8aaa84f49964ee7ba7ab24