Assign unique encap indices to loopback IPs while ensuring same order
Summary:
Our vendor SAI impl currently does not support
sharing same encap index value even when egress port and mac
are the same. We are working with vendors to do so, but
meanwhile the only thing we need is for the different
DSF nodes to agree on the IP->encap index value for
all nodes in the system. To do so, just sort the
loopback ips and assign encap indices in sorted
order.

Reviewed By: shri-khare

Differential Revision:
D41167094

LaMa Project: L1125642

fbshipit-source-id: 63091241b5fd514ae3f4e811cefe7a8a14d8d8d0