Assign encap id to neighbor entry where ASICs permit this
Summary:
For VOQ switches, we need a encap id for neighbor forwarding
to work. We can either have the SDK assign this or manage this ourselves.
Having SDK assign it means we have to bubble this up from HwSwitch back
upto SwSwitch post programming. Instead we choose to manage this assignment
ourselves. To this end, when programming a neighbor entry get the
next available encap idx ad assign it to neighbor entry. Note that
we do this in context of the update thread, so the updates SwitchState
always reflects the already allocated encap ids.

Reviewed By: nivinl

Differential Revision: D41256771

fbshipit-source-id: 1c32dc6c89b5b7daf941009a47dc784756e55436