Fail command if required subcommand is not passed
Summary:
Right now CLI silently exists with status 0 in cases where we have a passthrough command. This changes to at least fail the CLI with a generic message
```
A subcommand is required
Run with --help for more information.
```
Haven't figured out (if even possible) a CLI11 option to dump the possible subcommands to try, but this is good step in the right direction

Reviewed By: zechengh09

Differential Revision: D41346727

fbshipit-source-id: f7f3448565b17213b0ee33bad22ff4041f0d40c5