providing default resolver for port queue fields
Summary:
as titled, this is needed because thrift list node has no mechanism of overriding element type.

instead giving default resolver to move fast.

in future, we can provide thrift list node a mechanism to override an element type. but at this moment it is unnecessary.

Reviewed By: jasmeetbagga

Differential Revision:
D41298502

LaMa Project: L1125642

fbshipit-source-id: 05e33372204fdfe1bcaa2d181dacc06c1c966bb4