Make sure to call const api in getForwardingState
Summary:
This is a pretty annoying gotcha with thrift_cow apis. In getForwardingState `portToFwdState` here is a `const shared_ptr<T>&`, but T itself is not `const`. So calling `find` here ends up calling the non-const version of find which uses writableFields https://fburl.com/code/sd79pwo6. The side-effect is if this get method is every called on a published state we ended up crashing, even though we're not actually trying to modify state P555325689. Same bug lives in getPartnerState

For now forcing the obj to be const so we pick the correct find, but we need to figure out a way to return something like `const shared_ptr<const T>&` from all `cref` apis so we can catch this at compile time

Reviewed By: zechengh09

Differential Revision:
D41407075

LaMa Project: L1125642

fbshipit-source-id: 3950a44bcf785c88698af2b8c05d468f85dda349