Make firstVlanID() return an optional value
Summary:
We have configs which has pure l3 ports, adjust firstVlan() to return optional value
to make sure its accounted for in the existing flows. As part of this, introduced a
new API getFirstInterfaceMac() to get the first interface MAC from config / state,
which avoids having to get the vlan first and then get the MAC.

Marked as TODO a few occurrence where VLAN id is used to further build the
packet needed, will need to migrate this code to possibly work with / without
VLAN tag.

Reviewed By: jasmeetbagga

Differential Revision: D41105050

fbshipit-source-id: e7670fa350afe566136888dfb5fefcc9c7c1a03c