Recommit D39976852 with the crash fixed
Summary: Recommit D39976852 with the crash fixed. The crash was originating because state.aclMap() and state.aclTableGroupMap() are always non null with empty map. This causes ACLs to be not processed and acltablegroup to be processed. And since map is empty, fields.acl was populated with a null ptr when moving from preprod to trunk

Reviewed By: zechengh09

Differential Revision:
D41472777

LaMa Project: L1125642

fbshipit-source-id: 77c71e75bc43ebe55c90e5d064f878019367360b