Assert for encap index in neighbor state changes
Summary:
As titled. For Mock switch we have the feature
RESERVED_ENCAP_INDEX set to true. So we will assign encap
indices to nbr entries in UTs. Assert that.

TODO: For now the code asserts for encap indices on
both vlan and rif neighbors. As we refine the NeighborCacheImpl
code to do so only for VOQ switches + rif neighbors we can
set the checkEncapIndices parameter appropriately

Reviewed By: nivinl

Differential Revision:
D41420189

LaMa Project: L1125642

fbshipit-source-id: 3d9a6eedc950b6bcc7612818b83bad6ac0a0f72b