Generalize WaitForNeighbor* helpers to deal with neighbors on vlans or intfs
Summary:
For VOQ switches, neighbor tables will live on Interface objects.
So these UT helpers need to handle tables being on either vlan or
intfs. Gradually we will migrate all neighbor tables to be on intfs but
in the interim we need to test for either case

Reviewed By: nivinl

Differential Revision:
D41387823

LaMa Project: L1125642

fbshipit-source-id: 5203e99015b9a70db63660652b8ca1eff0988ceb