Initialize dsfNodes
Summary: This will keep the cfgr and fbcode copy in sync. If we don't initialize it, we need to explicitly set it for fsw, ssw, fa etc generated configs for production/lab explicitly . else it leads to build failures such as P561754674. Its simpler to initialize to empty map instead of making sweeping changes in coop templates in cfgr.

Reviewed By: jasmeetbagga

Differential Revision:
D41573946

LaMa Project: L1123871

fbshipit-source-id: bdccdf6f498bbfdf4530725eaebfff6b326e2ed1