reference wrapper for thrift cow objects
Summary:
In nested (or complex), data structures we have risk of invoking API with "writableField" in picture on published objects, causing crash.

This is because "const std::shared_ptr<T>" and dereferencing it with * operator leads to T and not const T.

We hit this many times as we moved to switch state to thrift cow but so far, used std::as_const to get around this.

This is fragile as it requires person using thrift cow to know, NOT to call non-const api on published object.

Providing an abstraction for ReferenceWrapper. Dereferencing RefererenceWrapper to reference of type const std::shared_ptr<T>& or const std::optional<T>& yields const T& (or const T*). On non-const shared_ptr or non-const optional, yields T& (or T*).

Methods such as cref or ref will return these objects.

Reviewed By: aeckert

Differential Revision:
D41560618

LaMa Project: L1125642

fbshipit-source-id: fdb0745c92d6bfef3e5dec98ee8e3aace79fa213