Handle PhyStats and PhyState fields in FBOSS2 CLI
Summary: PhyInfo now has 'state' and 'stats' field and we are separating the current PhyInfo into state and stats. This diff makes changes to the "show interface phy" CLI and uses the state and stats field if they exist, or the legacy fields if they don't. Plan is to deprecate the legacy fields once the changes are rolled out everywhere, and then will deprecate the corresponding CLI code too.

Reviewed By: rajank7

Differential Revision:
D41711385

LaMa Project: L1125642

fbshipit-source-id: 8bb58e69d179565f953f4891bc5ae74eb649e8b0