Start adding DsfSubscriber tests + bug fix
Summary:
As titled. Start adding UTs for DsfSubscriber code + fix
a bug which was revealed in e2e testing. In that we scheduled a
lambda which captured stack parameters by reference. During e2e
testing we solved this by making the state update blocking. However
a better solution is to just copy the shared ptrs and keep the
function async. This allows for better coalescing of updates

Reviewed By: shri-khare

Differential Revision:
D41826687

LaMa Project: L1125642

fbshipit-source-id: 9241f306f7fa55ea450622ce9e01b4c1032d4f01