Start populating PhyState and PhyStats in PhyInfo for SaiSwitch
Summary:
We currently combine phy state and stats in the same structs inside PhyInfo (for example PmdInfo stores both state and stats), which makes it hard to publish these entities separately to FSDB. Therefore, we want to separate state and stats structs. We'll need to store duplicate data for some time until clients like NGT and CLIs can migrate to using the new fields. The plan is to
1. Separate PhyState and PhyStats in SaiSwitch
2. Separate PhyState and PhyStats in BcmPort
3. Separate PhyState and PhyStats for Yamp XPHY
4. Separate PhyState and PhyStats for Minipack and Minipack2 XPHY
5. Point all clients to PhyState and PhyStats rather than PhyInfo
6. Deprecate old fields in PhyInfo

This diff starts implementing 1) and adds PhyState and PhyStats to PhyInfo for SaiSwitch. Currently, the diff only publishes FEC counters. Other state and stats are added in next diffs.

Reviewed By: jasmeetbagga

Differential Revision:
D41700734

Privacy Context Container: L1125642

fbshipit-source-id: 158fc7f9d7e8e9f9d5a96307ec6c8bb77c6f5754