HW test for validating PhyInfo
Summary:
Added a check as part of the PortProfileTests so that PhyInfo gets validated for all speeds. This test can verify that the expected attributes have been read correctly and populated in PhyInfo structs. Only validating the new state and stats fields since those are the ones that will remain going forward and others will be deprecated. Note that the correctness of the values of these fields cannot be validated without link up in normal mode (not loopback), so leaving that part of verification for link test.

I had to add 2 helper functions in HwSwitch to return a flag to indicate that the SDK supports the diagnostics. This is because the SDK flags aren't propagated to the tests. I also didn't want to check for HwAsic::Feature in the test, because then we miss verifying these diagnostics for future new ASICs because we also disable these features initially.

Reviewed By: jasmeetbagga

Differential Revision:
D41970338

Privacy Context Container: L1125642

fbshipit-source-id: cd387d35a02fb93663383089ada6599ccb30952e