Handle feature rollback for SW MAC learning
Summary:
Original issue:
To mitigate SEV S289408, SW stack rollback was done. As part of this
rollback, features that got enabled during the push also got
rolled back. So, SW MAC learning feature that got enabled during the push
was disabled. But for this, FBOSS agent didn't deregister the L2 learn
callbacks in some ASIC platforms causing SDK to keep sending the
L2 learn callbacks. Also, the feature state was out of sync
between agent and SDK/ASIC.

Due to this, when the switch (non TH ASICs) got warmbooted next time, it
resulted in agent crashes in some of the switches (crash depends on timing)
as reported in https://fburl.com/6d316azl.

Root cause and fix:
The reason why L2 learn callback was not deregistered when the SW MAC
learning got disabled is - the warmboot cache that holds the SDK/ASIC
state was incorrectly updated as SW MAC learning was already disabled in
SDK/ASIC (the logic that popoulates this state for L2 learning mode works
only for TH ASIC and not for other ASICs). So, FBOSS agent skipped to
deregister the L2 learn callback in non TH ASICs. The fix is to ignore
the warmboot cache state so that L2 learn callback would be deregistered
based on the agent state. This is already done when SW MAC learning is
enabled and extending this behavior now for SW MAC learning disabled case also.

Also, need to check with Broadcom on how to identify the L2 learning mode in
SDK/ASIC that would work for all ASIC platforms.

Reviewed By: daiwei1983

Differential Revision: D42283734

fbshipit-source-id: d813e851240e153c862f576a0d153bfe0ea947b2