Relax acl pkt counter check in voq tests
Summary:
Assert for acl counter getting incremented.
In HwVoqSwitchTest.sendPacket* tests I see ACL counter
incrementing by 2 not 1. Likely this is because the
looped packet is also counted before getting dropped.
So relax the check to assert for GT. Why not just check
for +2. This is to allow for different asic behaviors
where looped packet is not counted before drop.

Reviewed By: nivinl

Differential Revision: D42446006

fbshipit-source-id: 0d52c2d1b66cb78c915dd8c013dfe246e9d27be5