respect TraverseHelper::shouldShortCircuit() during DeltaVisitor traversal
Summary:
This is the main interface for influencing traverse behavior
for now. DeltaVisitor will not recurse into a child if
shouldShortCircuit() returns true.

Note that this happens BEFORE we visit a given node for DeltaVisitor,
so we need to take care to make sure we don't short-circuit
prematurely if there is a reason we need to visit the current node.

Reviewed By: peygar

Differential Revision: D42089227

fbshipit-source-id: 6b76b95fe6c4230f6b010ff66a3e8cc350852830