add UNPUBLISHED mode to RecurseVisitor
Summary:
This adds a new mode to RecurseVisitor to only visit
UNPUBLISHED nodes. I had hoped to be able to do this in a
TraverseHelper, but unfortunately this would need access to the
underlying node when making a short circuit decision, which will
introduce a lot of complexity in to TraverseHelper. The other option
was potentially to support a return type in the visitor to signal we
should short circuit too, but ultimately I decided this was simpler
for this one use case.

The plan here is be able to create a visitor that publishes the tree,
but also ensures that we have added all new paths to the
SubscriptionPathStore to resolve extended subscriptions. Right now,
adding new paths to SubscriptionPathStore to determine if any extended
subscriptions match happens in the main serveSubscriptions visit
function. The issue is that this gets in the way of our short circuit
logic because even if don't have any subscriptions yet, we might get
new ones as we recurse further. Pulling processAddedPath() calls out
of the main serveSubscriptions visitor will unlock ability to
shortcircuit traversal.

Reviewed By: peygar

Differential Revision: D42154020

fbshipit-source-id: 1fecb03c990758f0b79d3511c145602b8aca7be5