introduce TraverseHelper interface
Summary:
This introduces a new interface to influence how we traverse
in a visitor. This is intended to be subclassed and stateful to allow
using external state to influence how we traverse.

In the context of fsdb, we want to let the set of active subscriptions
influence our traversal in DeltaVisitor + RecurseVisitor.

Reviewed By: peygar

Differential Revision: D42089228

fbshipit-source-id: aeacf27ec1b71e96359f4619871e7ab161a0751d