allow signaling which visitor is active to TraverseHelper
Summary:
For the TraverseHelper we will create for short circuiting
during DeltaVisitor FULL visits, we need to influence both the
DeltaVisitor and the RecurseVisitor it creates on fully added/removed
nodes.

However, the short circuit logic won't be exactly the same in the two
stages, so this adds a way to pass in which visitor is currently
active to the TraverseHelper::shouldShortCircuit() function.

Reviewed By: peygar

Differential Revision: D42106991

fbshipit-source-id: cf9a70db12f7da019221ee8dd7c96ec60c2cac7f