add ability to control order in RecurseVisitor
Summary:
Another nit, but we can't easily publish a parent node before
a child node in our publishAndAdd visitor we are creating. This adds a
new arg for whether to visit as we recurse or do a depth first recurse
and visit leaves first.

This will allow us to publish all children before we publish a parent
node.

Reviewed By: peygar

Differential Revision: D42155178

fbshipit-source-id: 371687292e8031715b781d3535fdb6c8663c1529