allow passing traverser to visitor fn
Summary:
This adds logic for both DeltaVisitor and RecurseVisitor to
pass the TraverseHelper to the visit function rather than the path,
depending on the visit function signature.

This will allow us to also leverage the state of the TraverseHelper in
visit functions to speed them up. In our case, the TraverseHelper we
create for serving subscriptions will already have the
SubscriptionPathStore* at a path, so we can speed up finding path
stores and relevant subscriptions later on.

Reviewed By: peygar

Differential Revision: D42154019

fbshipit-source-id: c2bf94ec8fb9f7f83c08bf02efc63c74e59ecfd5