centralize deltavisitor logic to skip traversing nodes where pointers match
Summary:
Nit, but we had logic that skips if pointers are equal in
different type specific field visit functions.

Rather than having these in the different field visitors, let's just
centralize it in one place in visitNode. Thought this was clearer as
we add more reasons we might not continue recursing w/ the
TraverseHelper stuff.

Reviewed By: peygar

Differential Revision: D42106994

fbshipit-source-id: 7a00bece610319906e1a71ab30a530235aa179f2