more logging around reading warm boot flags
Summary:
We're seeing a case where we expected a cold boot due to the
force cold boot flag, but aren't sure if the tool that dropped that
flag had an issue.

This is a minor helper to add logging when looking for these flags so
we can tell what agent saw when making a decision for whether to warm
boot or not.

Differential Revision:
D42485197

Privacy Context Container: L1125642

fbshipit-source-id: 7f0d148d1b8c1b53fe983849bb27370922ae5b9b