Propagate needResetDataPath flag in qsfp service state machine to programExternalPhyPorts
Summary:
There is a flag needResetDataPath inside QSFP Service state machine which is set to true whenever agent cold boots. This flag is then used for resetting the data path of Transceivers. The idea is that when agent cold boots, we need to reset the data path of optics since the outgoing signal from ASIC changes.

We need to follow similar logic when there is an XPHY in the system. When agent cold boots, we should reinit the XPHY and then reinit transceivers. This diff stack adds support for that.

This diff - Propagates needResetDataPath to programExternalPhyPorts function
Next diff - Propagate needResetDataPath to programOnePort
Diff after - Use needResetDataPath in programOnePort

Reviewed By: rajank7, birdsoup

Differential Revision:
D42455197

Privacy Context Container: L1125642

fbshipit-source-id: 20a2bf9be5c99670757e11d75c18a43c6e21bee6