Skip hw writes and cache switch state in DsfSubscriber
Summary:
In FsdbAgentDsfSubscriberTests we get DSF subcriber
to subscribe to local node's state. This is a handy way to
test DsfSubsriber<->FSDB interaction, but it runs afoul
when we try to reapply local state (sysports, rifs) as
remote objects. So add a flag to allow skipping HW writes.
Also add another flag to cache updated switch state on
DsfSubscriber so we can query it in tests

Reviewed By: nivinl

Differential Revision:
D42495262

Privacy Context Container: L1125642

fbshipit-source-id: 1581643576ecdd58fba892243baab319ba67e9dc