Fix HwVoqSwitchTest.collectStats failure - Avoid reading FEC counters twice on Fabric Ports
Summary:
Looks like we are currently reading the FEC counters twice for Fabric ports. One place is the one in this diff, the 2nd place in switch/SaiPortManager.

The 1st place was probably added initially when SDK didn't support reading FEC counters on all kinds of ports.

The 2nd place is actually the correct place to read since it only tries to read when FEC is supported.
```
  handle->port->updateStats(supportedStats(portId), SAI_STATS_MODE_READ);
  if (fecStatsSupported(portId)) {
    handle->port->updateStats(
        {SAI_PORT_STAT_IF_IN_FEC_CORRECTABLE_FRAMES,
         SAI_PORT_STAT_IF_IN_FEC_NOT_CORRECTABLE_FRAMES},
        SAI_STATS_MODE_READ_AND_CLEAR);
  }
```

This duplicate read was the cause of HwVoqSwitchTest.collectStats failure that we started seeing after D42391600 because with D42391600 now SDK returns SAI_STATS_UNSUPPORTED when we read FEC counters on a FEC disabled port.

Differential Revision:
D42527555

Privacy Context Container: L1125642

fbshipit-source-id: 73e068e18ebe386e20fec5dd52cdd52ccf06ab23