fix hw macsec tests
Summary:
fixing qsfp test failures introduced by D41791960. prior to migration of port/port map to thrift cow rxSak was represented as std::map but as port moved to thrift cow, it came to be represented as std::vector, .

Restoring use of this data structure to keep the behavior same.

Reviewed By: zechengh09

Differential Revision: D42613199

fbshipit-source-id: d348fb2bc11beb7d1de875ffab5d54b8bca21690