agent ensemble
Summary:
something on the lines of hw switch ensemble.

the purpose of this is to start and stop agent, and provide methods which hide details which otherwise would bloat test code. provide most commonly used and needed methods which can be used in tests or benchmarks working at an agent level

Reviewed By: jasmeetbagga

Differential Revision: D42425625

fbshipit-source-id: 98d7c9e14e0267d5bb4d8be0ac200274683e58b5