Update getCurrentStateJSON to work with thrift path
Summary:
As thrift state is converted to thrift cow, update getCurrentStateJSON to work with thrift path. We'll use path visitors to walk down switch state (which is similar to how we did it in json).

Will follow up with changes in fboss2 CLI which relies on this thrift call.

Reviewed By: jasmeetbagga

Differential Revision:
D42669464

Privacy Context Container: L1125642

fbshipit-source-id: 0de1e07487c3fa5aac34d4f9ad5c1d48fb0a8326