Set isLocal to false before programming remote neighbors
Summary:
A neighbor is local only to the DSF Node the neighbor was resolved on, but it
is remote for every other DSF Node.

When a DSF node publishes its neighbor entries, that carries isLocal = True.
Every DSF node subscriber needs to set isLocal = False before programming it.
This diff performs that.

Reviewed By: jasmeetbagga

Differential Revision:
D42493215

Privacy Context Container: L1125642

fbshipit-source-id: 5cef101216a12e8bb0374b1b59023ff6f71a92d9