Set interface in configs for vlans
Summary:
Stacked diff populates interfaceIDs for a port in the switch state.
For non-VOQ switches, this is implemented as follows:
  - determine vlan to interface mapping from config
  - for every port in config:
      find vlan the port belongs to
      lookup vlan to interface mapping
      use it to populate interface for that port

Thus, we need all the above fields populated in the config.
These fields are always populated in the production config, but some unit tests
didn't always populate these fields correctly, so fix it.'

Reviewed By: jasmeetbagga

Differential Revision:
D42606086

Privacy Context Container: L1125642

fbshipit-source-id: 31ad62c85c3ec87036ecf84b17b88344de7718cf