IPv6Handler: getSwitchIntfIPv6 instead of getSwitchVlanIPv6
Summary:
VOQ/Fabric switches require that the packets are not tagged with VLAN.
This diff is part of the diff stack that enhances wedge_agent to deal with VLAN
tagged as well as untagged packets.

As titled, so it works with NPU as well as VOQ switches.

Reviewed By: jasmeetbagga

Differential Revision:
D42699145

Privacy Context Container: L1125642

fbshipit-source-id: bad8a8a4fbcb9ed718aeff75108eecdd773b41c8