Add a check for interface id falling within sys port range
Summary:
As titled. With this I had to also get rid of a
couple of test which were changing switch id, type on
the fly w/o updating other aspects of the config. We could
make those work with appropriate changes to test configs,
but since we never plan to change switch type, id
dynamically on a box, omit the tests

Reviewed By: nivinl

Differential Revision:
D43036629

Privacy Context Container: L1125642

fbshipit-source-id: 2947517e5b3f7a96e2c44eca55dcedfec9efdad8