Removed strick check for channels in getChannelFlags()
Summary:
We see from the log file that the CMIS optics goes for remediation where the application code is being programmed. For some time the optics returns all garbled data causing even the EEPROM checksum to fail. The random data returned by the optics causes its current application to be identified as wrong value and it starts referring to wrong channel id which causes the crash. When optics is in stable state then all its register values are correct but only during initialization time, its value are random.
In this diff we hav removed the strict check for the channel id in getChannelFlag and printing the error so that we will come to know what is causing this issue in future. This issue is rare to happen.

Reviewed By: harshitgulati18

Differential Revision: D43034243

fbshipit-source-id: 992d0174bca4f8ef51cc596938645ffa7a9e8ed7