add ttl/counter qualifiers for sdk 1.58.0/1.60.0
Summary:
As titled, add ttl/counter qualifiers and actions for SDK 1.58.0 and 1.60.0.

For Tajo, the first ACL table qualifier acts as the super set since it defines the key profile. For all ACL group tests, include TTL and Counter for SDK 1.58.0 and 1.60.0 so that subsequent tables can be created with TTL qualifier only.

Differential Revision: D43103366

fbshipit-source-id: 3647750f9d4ebad842e650f4b67cfd06acf4b2b8