Move PHY SDK capabilities to a separate file
Summary: The hw tests need to know if the SDKs support a certain PHY diagnostic or not so that it can then verify the actual support in software. Currently, we are functions in HwSwitch that tell whether these diagnostics are supported by SDK or not. However, these functions need not be in HwSwitch and can be in their standalone files. This diff adds PhyCapabilities.{h,cpp} and the next diff will remove the current functions in HwSwitch and make use of these capabilities

Reviewed By: jasmeetbagga

Differential Revision: D43222385

fbshipit-source-id: a07c47fbf9d18d2989964ff1ccf1a6bdc5c1d5f6