Ensure Validity of Next Hop IP Addresses
Summary: When ProdInvariantTests are run for prod configs with many downlink ports, it becomes very likely that the last octet of a next hop IP addresses exceeds 255, which fails existing checks in EcmpSetupHelper. To resolve this issue, the final value of lastOctet is computed by taking the mod 255 of lastOctet.

Reviewed By: zechengh09

Differential Revision: D43260346

fbshipit-source-id: adb1591b3ad333c119f933e0546b7a010b927f95